backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pandas>=2.0.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pillow>=10.0.0
